App now gets likes and time
The app gets all of the user's pictures posted times and the likes that
those pictures received. The time is also displayed in a readable
manner, rather than as a unix timestamp. Now I just need to make the
values graph and do some styling and I'll be able to wrap this up.